refactor: wallet feature
