--- src/features/wallet/common/router.ts
@@ -1,4 +1,0 @@
-import { paymentRouter } from "../payment/routers";
-import { walletV3Router } from "../routers";
-
-export const walletRouter = [walletV3Router, paymentRouter];
